split utils functions
