+++ src/constants/constants.js
@@ -1,0 +1,14 @@
+/* eslint-disable max-len */
+
+const ERRORS = {
+  MISSING_CASE:
+    '"toCase" query param is required. Correct request is: "/<TEXT_TO_CONVERT>?toCase=<CASE_NAME>".',
+  MISSING_TEXT:
+    'Text to convert is required. Correct request is: "/<TEXT_TO_CONVERT>?toCase=<CASE_NAME>".',
+  INVALID_CASE:
+    'This case is not supported. Available cases: SNAKE, KEBAB, CAMEL, PASCAL, UPPER.',
+};
+
+const CASES = ['SNAKE', 'KEBAB', 'CAMEL', 'PASCAL', 'UPPER'];
+
+module.exports = { ERRORS, CASES };
